Add files via upload
This is one of the data entry interface for Operation data entry. There are two section of the interface, first section is the data entry itself which users can enter the data by filling the text box with certain data. Second section is the database which users can see the data that they have entered. There's also manipulation button which is edit and delete, this allows users to do data modification.